Use #pragma once rather than your common header guards.
of header guards, but it's less code to type, avoids name clashes, and it improves
compilation speed slightly.  MSVC, g++, and clang++ all support #pragma once.  The likelihood
of the compiler failing to handle this seems low, so if any problems arise I will be reverting it
to your normal header guards, otherwise #pragma once stays.